[KDM!] Added tower defense technical prototype
Key Design Moment (KDM): I decided to take a slightly different approach for this project. In a previous series of technical tests, I learned the techniques to make a tower defense / missile command sort of game. I've just committed all those for the first time in this project in this commit. I altered the style a little bit by using a free-style tile placement on a 2D map instead of restricting the tower (or hero, in my version) placement in lanes. The freestyle placement, allowing the player to move a hero between level-specific tile choices around the player's base, is supposed to allow the player to  have more position moves in a 360 degree circle.