Added progress on Quest Idler
Going for the rough and done thing for this one. I kind of like the raw css look but it's just there temporarily. You can click on the avatar and level up indefinitely. Working on this one further today. Also added the rest of the idlers that will be in the carousel.