Overhaul typography and video game poetry ideas
This commit really makes it all look consistent and neat within the boundaries of neat that can be in this particular game. The poetry overhaul is especially important: I wanted to fit this project/game inside the digital or video poetry movement, continuing on with experimenting the installation of poetry inside video games. More emphasis or attention is called on the different manifestations of poetry itself (old English Shakespeare and post-modernist, fragmented styles), embedded inside the various contexts of this game.